feat: update tests on frame support pallet_instance

+++ substrate/frame/support/test/tests/pallet_instance.rs
@@ -339,7 +339,7 @@ impl pallet2::Config<pallet::Instance1> for Runtime {
 	type RuntimeEvent = RuntimeEvent;
 }
 
-pub type Header = sp_runtime::generic::Header<u32, sp_runtime::traits::BlakeTwo256>;
+pub type Header = sp_runtime::generic::Header<u64, sp_runtime::traits::BlakeTwo256>;
 pub type Block = sp_runtime::generic::Block<Header, UncheckedExtrinsic>;
 pub type UncheckedExtrinsic = sp_runtime::generic::UncheckedExtrinsic<u32, RuntimeCall, (), ()>;
